requirements.txt
beautifulsoup4==4.12.2
html5lib==1.1
pandas==2.1.2
PyYAML==6.0.1
requests==2.31.0
